+++ yalafi/packages/babel.py
@@ -3,6 +3,12 @@
 #   YaLafi module for LaTeX package babel
 #
 
+#   do the macros / environments always break the text flow?
+#
+foreignlang_break = False
+selectlang_break = True
+otherlang_break = False
+
 #   map between language codes for babel and LanguageTool
 #   - sorted according to xx-XX code
 #
@@ -54,7 +60,7 @@ language_map = {
     'chinese': 'zh-CN',
 }
 
-from yalafi.defs import InitModule, Macro, LanguageToken
+from yalafi.defs import InitModule, Macro, LanguageToken, Environ, MacroToken
 
 require_packages = []
 
@@ -65,18 +71,27 @@ def init_module(parser, options):
 
     macros_python = [
 
+        # the following is for \end{otherlanguage}
+        Macro(parms, '\\babel@skip@space', args='', repl=''),
         Macro(parms, '\\foreignlanguage', args='OAA', repl=h_foreignlanguage),
         Macro(parms, '\\selectlanguage', args='A', repl=h_selectlanguage),
 
     ]
 
-    environments = []
+    environments = [
+
+        Environ(parms, 'otherlanguage', args='A', repl=h_begin_otherlang,
+                                add_pars=False, end_func=h_end_otherlang),
+        Environ(parms, 'otherlanguage*', args='A', repl=h_begin_otherlang,
+                                add_pars=False, end_func=h_end_otherlang_star),
+
+    ]
 
     inject_tokens = []
     if options:
         # set current language to the last in option list
         inject_tokens = [LanguageToken(0, lang=translate_lang(options[-1][0]),
-                                                hard=True)]
+                                                hard=True, brk=True)]
 
     return InitModule(macros_latex=macros_latex, macros_python=macros_python,
                         environments=environments, inject_tokens=inject_tokens)
@@ -91,10 +106,21 @@ def translate_lang(lang):
 
 def h_foreignlanguage(parser, buf, mac, args, pos):
     lang = translate_lang(parser.get_text_expanded(args[1]).strip())
-    return ([LanguageToken(pos, lang=lang)] + args[2]
+    return ([LanguageToken(pos, lang=lang, brk=foreignlang_break)] + args[2]
                         + [LanguageToken(args[2][-1].pos, back=True)])
 
 def h_selectlanguage(parser, buf, mac, args, pos):
     lang = translate_lang(parser.get_text_expanded(args[0]).strip())
-    return [LanguageToken(pos, lang=lang, hard=True)]
+    return [LanguageToken(pos, lang=lang, hard=True, brk=selectlang_break)]
+
+def h_begin_otherlang(parser, buf, mac, args, pos):
+    lang = translate_lang(parser.get_text_expanded(args[0]).strip())
+    return [LanguageToken(pos, lang=lang, brk=otherlang_break)]
+
+def h_end_otherlang(parser, buf, mac, args, pos):
+    return [LanguageToken(pos, back=True),
+                MacroToken(pos, '\\babel@skip@space')]
+
+def h_end_otherlang_star(parser, buf, mac, args, pos):
+    return [LanguageToken(pos, back=True)]